requirements.txt
setuptools
wheel
ipython
Flask
flask_login
flask_wtf
flask_sqlalchemy
flask_migrate
flask_login
flask_mail
flask_bootstrap
flask_moment
flask_babel
elasticsearch
redis
rq
python-dotenv
email_validator
click
faker